Print current gen every 100 gens

+++ tensorneat/src/tensorneat/pipeline.py
@@ -171,10 +171,11 @@ class Pipeline(StatefulBaseClass):
 
                 # overwrite the previous progress line instead of appending
                 # move to file start, write the line, then truncate the rest
-                f_log.seek(0)
-                f_log.write(f"Generation {gen+1}/{self.generation_limit}\n")
-                f_log.truncate()
-                f_log.flush()
+                if gen%100 == 0:
+                    f_log.seek(0)
+                    f_log.write(f"Generation {gen+1}/{self.generation_limit}\n")
+                    f_log.truncate()
+                    f_log.flush()
 
         if int(state.generation) >= self.generation_limit:
             print("Generation limit reached!")
